Все базовые функции работают нормально

+++ run.js
@@ -24,21 +24,31 @@ DiscordClient.on('ready', () => {
 })
 
 DiscordClient.on('message', msg => {
-    if (msg.content.match(/^!пидордня/) || msg.content.match(/^!пидорня/)) {
-        participantsRepository
-            .IsParticipantExists(msg.author.id, msg.guild.id)
-            .then(isExists => {
-                if (isExists) {
-                    ChatFunctions.temporaryMessage(msg.channel, "You're already participating in this game, silly", 7000);
-                } else {
-                    participantsRepository.AddParticipant(msg.author.id, msg.guild.id, ChatFunctions.getNickname(msg));
-                    ChatFunctions.temporaryMessage(msg.channel, "Alright, you're in, " + ChatFunctions.getNickname(msg), 5000)
-                }
-            });
+  if (msg.author.bot) return;
+  if (msg.channel.type == "dm") return;
 
-        ChatFunctions.deleteMessage(msg, 2000);
-        return;
-    }
+  /*if (msg.content.match(/^!удалиэто/)) {
+    console.log(`команда удалиэто`);
+    const response = msg.channel.send("сообщение с командой удалено");
+    ChatFunctions.deleteMessage(msg, 2000);
+  }*/
+
+  if (msg.content.match(/^!пидордня/) || msg.content.match(/^!пидорня/)) {
+    participantsRepository
+      .IsParticipantExists(msg.author.id, msg.guild.id)
+        .then(isExists => {
+          if (isExists) {
+            const response = msg.channel.send("Ты уже участвуешь в этой игре, чел");
+            //ChatFunctions.temporaryMessage(msg.channel, "You're already participating in this game, silly", 7000);
+          } else {
+            participantsRepository.AddParticipant(msg.author.id, msg.guild.id, ChatFunctions.getNickname(msg));
+            const response = msg.channel.send("Окей, ты в игре, " + ChatFunctions.getNickname(msg));
+            //ChatFunctions.temporaryMessage(msg.channel, "Alright, you're in, " + ChatFunctions.getNickname(msg), 5000)
+          }
+        });
+    ChatFunctions.deleteMessage(msg, 2000);
+    return;
+  }
 
     if (msg.content.match(/^!ктопидор/)) {
         game.CanStartGame(msg.guild.id).then(() => {
@@ -46,7 +56,8 @@ DiscordClient.on('message', msg => {
                 await game.Tease(msg.channel).then();
                 msg.channel.send(winMsg);
             }, reject => {
-                ChatFunctions.temporaryMessage(msg.channel, reject, 8000);
+                msg.channel.send(reject);
+                //ChatFunctions.temporaryMessage(msg.channel, reject, 8000);
             });
         }, reject => {
             msg.channel.send("А пидор сегодня - " + reject);
@@ -55,26 +66,29 @@ DiscordClient.on('message', msg => {
         ChatFunctions.deleteMessage(msg, 1000);
     }
 
-    if (msg.content.match(/^!топпидоров/)) {
+    if (msg.content.match(/^!топпидоров/) || msg.content.match(/^!топ/)) {
         game.GetStats(msg.guild.id).then(message => {
-            ChatFunctions.temporaryMessage(msg.channel, message, 15000);
+            msg.channel.send(message);
+            //ChatFunctions.temporaryMessage(msg.channel, message, 15000);
         });
         ChatFunctions.deleteMessage(msg, 1000);
         return;
     }
 
     if (msg.content.match(/^!исключить/)) {
-        let chunks = msg.message.split(' ');
-        chunks.splice(0, 1);
-        let discordId = chunks.join('');
+        /*let chunks = msg.content.slice(prefix.length).trim().split(/ +/g);
+        if(chunks[1] != undefined) {
 
-        if (msg.author.id !== '207169330549358592') {
-            ChatFunctions.temporaryMessage(msg.channel, "Вы кто такой? Я вас не звал. Идите нахуй!");
-        } else {
-            ChatFunctions.temporaryMessage(msg.channel, "Пидарнул пидорка нахуй");
-        }
-        ChatFunctions.deleteMessage(msg, 3000);
-        participantsRepository.RemoveParticipant(discordId, msg.guild.id)
+        }*/
+        let discordId = msg.mentions.members.first().id;
+        //let chunks = msg.content.slice(prefix.length).trim().split(/ +/g);
+        //let chunks = msg.message.split(' ');
+        //chunks.splice(0, 1);
+        //let discordId = chunks[0];
+
+        msg.channel.send("Ок");
+
+        participantsRepository.RemoveParticipant(discordId, msg.guild.id);
         return;
     }
 });
